LayoutTests/inspector/timeline/timeline-recalculate-styles.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>

<style>
.test-style {
    color: red;
}
</style>

<script>

function performActions()
{
    var element = document.createElement("div");
    element.className = "test-style";
    element.innerHTML = "Test data";
    document.body.appendChild(element);
}

function test()
{
    InspectorTest.performActionsAndPrint("performActions()", "RecalculateStyles");
}

</script>
</head>

<body onload="runTest()">
<p>
Tests the Timeline API instrumentation of a style recalculation event
</p>

</body>
</html>
